pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cachetools>=5.5.1",
    "fastapi>=0.115.8",
    "pyjwt>=2.10.1",
    "uvicorn>=0.34.0",
]
